+++ Enemy.cpp
@@ -5,7 +5,7 @@
 
 Enemy::Enemy() : Enemy(0, sf::Vector2(0.f, 0.f), 0, 0, nullptr, 0) {}
 
-Enemy::Enemy(int r, sf::Vector2f position, int speed, int health, Weapon *weapon, int xp)
+Enemy::Enemy(int r, sf::Vector2f position, float speed, int health, Weapon *weapon, int xp)
     : Character(r, position, sf::Color::Green, speed, false, health, weapon),
       xp(xp) {}
