+++ Skeleton.h
@@ -8,6 +8,6 @@ class Skeleton : public Enemy
 {
 public:
     Skeleton();
-    Skeleton(int r, sf::Vector2f position, int speed, int health, Weapon *weapon, int xp);
+    Skeleton(int r, sf::Vector2f position, float speed, int health, Weapon *weapon, int xp);
     std::optional<Projectile> updateAI(sf::Vector2f playerPosition) override;
 };
